requirements.txt
absl-py==2.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
attrs==25.3.0
av==12.3.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.8.2
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
contourpy==1.1.1
ctranslate2==4.5.0
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
email-validator==2.3.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.124.4
faster-whisper==1.1.0
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==25.2.10
folium==0.18.0
fonttools==4.57.0
fsspec==2025.3.0
gast==0.4.0
google-ai-generativelanguage==0.1.0
google-api-core==2.29.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-generativeai==0.1.0rc1
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.11.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
jax==0.4.13
Jinja2==3.1.6
jiter==0.9.1
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
looseversion==1.3.0
lxml==6.0.2
Markdown==3.7
markup==0.2.2
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.101.0
opencv-contrib-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.0.3
pathlib==1.0.1
pillow==10.4.0
proto-plus==1.27.1
protobuf==4.25.8
prov==2.0.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.1
pymongo==4.10.1
pyodbc==5.2.0
pyparsing==3.1.4
pypdf==5.9.0
pyproj==3.5.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
python-tds==1.16.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
reportlab==4.4.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==3.2.1
simple-websocket==1.1.0
simpleaudio==1.0.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.30
sqlalchemy_pytds==1.0.2
starlette==0.44.0
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
traits==6.3.2
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.127
ultralytics-thop==2.0.14
urllib3==2.2.3
uvicorn==0.33.0
vision==1.0.0
Werkzeug==3.0.2
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.1.2
xyzservices==2025.4.0
zipp==3.20.2
